+++ appyx-components/experimental/cards/android/src/main/kotlin/com/bumble/appyx/components/experimental/cards/android/DatingCards.kt
@@ -11,13 +11,13 @@ import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.unit.dp
 import com.bumble.appyx.components.experimental.cards.Cards
 import com.bumble.appyx.components.experimental.cards.CardsModel
-import com.bumble.appyx.components.experimental.cards.ui.CardsMotionController
-import com.bumble.appyx.interactions.core.AppyxComponent
-import com.bumble.appyx.interactions.core.gesture.GestureValidator.Companion.permissiveValidator
-import com.bumble.appyx.interactions.core.ui.helper.AppyxComponentSetup
-import com.bumble.appyx.interactions.theme.appyx_dark
-import com.bumble.appyx.samples.common.profile.Profile
-import com.bumble.appyx.samples.common.profile.ProfileCard
+import com.bumble.appyx.components.experimental.cards.ui.CardsVisualisation
+import com.bumble.appyx.demos.common.profile.Profile
+import com.bumble.appyx.demos.common.profile.ProfileCard
+import com.bumble.appyx.interactions.composable.AppyxInteractionsContainer
+import com.bumble.appyx.interactions.gesture.GestureValidator.Companion.permissiveValidator
+import com.bumble.appyx.interactions.ui.helper.AppyxComponentSetup
+import com.bumble.appyx.interactions.utils.ui.theme.appyx_dark
 import kotlin.math.roundToInt
 
 @Composable
@@ -30,15 +30,15 @@ fun DatingCards(modifier: Modifier = Modifier) {
                 },
                 savedStateMap = null
             ),
-            motionController = { CardsMotionController(it) },
-            gestureFactory = { CardsMotionController.Gestures(it) },
+            visualisation = { CardsVisualisation(it) },
+            gestureFactory = { CardsVisualisation.Gestures(it) },
             animateSettle = true
         )
     }
 
     AppyxComponentSetup(cards)
 
-    AppyxComponent(
+    AppyxInteractionsContainer(
         modifier = modifier
             .fillMaxSize()
             .background(appyx_dark)
@@ -47,10 +47,11 @@ fun DatingCards(modifier: Modifier = Modifier) {
         screenHeightPx = (LocalConfiguration.current.screenHeightDp * LocalDensity.current.density).roundToInt(),
         appyxComponent = cards,
         gestureValidator = permissiveValidator,
-    ) { elementUiModel ->
+    ) { element ->
         ProfileCard(
-            profile = elementUiModel.element.interactionTarget.profile,
-            modifier = Modifier.fillMaxSize().then(elementUiModel.modifier)
+            profile = element.interactionTarget.profile,
+            modifier = Modifier
+                .fillMaxSize()
         )
     }
 }