+++ appyx-components/experimental/promoter/common/src/commonMain/kotlin/com/bumble/appyx/components/experimental/promoter/ui/TargetUiState.kt
@@ -1,15 +1,17 @@
 package com.bumble.appyx.components.experimental.promoter.ui
 
-import com.bumble.appyx.interactions.core.ui.property.impl.AngularPosition
-import com.bumble.appyx.interactions.core.ui.property.impl.RotationY
-import com.bumble.appyx.interactions.core.ui.property.impl.RotationZ
-import com.bumble.appyx.interactions.core.ui.property.impl.Scale
-import com.bumble.appyx.interactions.core.ui.property.impl.position.PositionInside
-import com.bumble.appyx.interactions.core.ui.state.MutableUiStateSpecs
+import com.bumble.appyx.interactions.ui.property.impl.AngularPosition
+import com.bumble.appyx.interactions.ui.property.impl.RotationY
+import com.bumble.appyx.interactions.ui.property.impl.RotationZ
+import com.bumble.appyx.interactions.ui.property.impl.Scale
+import com.bumble.appyx.interactions.ui.property.impl.position.PositionAlignment
+import com.bumble.appyx.interactions.ui.property.impl.position.PositionOffset
+import com.bumble.appyx.interactions.ui.state.MutableUiStateSpecs
 
 @MutableUiStateSpecs
 class TargetUiState(
-    val position: PositionInside.Target,
+    val positionAlignment: PositionAlignment.Target,
+    val positionOffset: PositionOffset.Target,
     val angularPosition: AngularPosition.Target,
     val scale: Scale.Target,
     val rotationY: RotationY.Target,