+++ appyx-components/experimental/puzzle15/web/src/wasmJsMain/kotlin/com/bumble/appyx/experimental/puzzle15/web/main.js.kt
@@ -1,0 +1,51 @@
+package com.bumble.appyx.experimental.puzzle15.web
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.focusable
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.Surface
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.ExperimentalComposeUiApi
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.focus.FocusRequester
+import androidx.compose.ui.focus.focusRequester
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.onSizeChanged
+import androidx.compose.ui.unit.IntSize
+import androidx.compose.ui.window.CanvasBasedWindow
+import com.bumble.appyx.components.experimental.puzzle15.ui.Puzzle15Ui
+
+external fun onWasmReady(onReady: () -> Unit)
+
+@OptIn(ExperimentalComposeUiApi::class)
+fun main() {
+    onWasmReady {
+        CanvasBasedWindow("Puzzle15") {
+            val requester = remember { FocusRequester() }
+            var size by remember { mutableStateOf(IntSize.Zero) }
+            Surface(
+                modifier = Modifier
+                    .fillMaxSize()
+                    .onSizeChanged { size = it }
+            ) {
+                Puzzle15Ui(
+                    screenWidthPx = size.width,
+                    screenHeightPx = size.height,
+                    modifier = Modifier
+                        .fillMaxSize()
+                        .background(Color.Black)
+                        .focusRequester(requester)
+                        .focusable(),
+                )
+            }
+
+            LaunchedEffect(Unit) {
+                requester.requestFocus()
+            }
+        }
+    }
+}