+++ appyx-components/internal/test-drive/android/src/androidTest/kotlin/com/bumble/appyx/components/internal/testdrive/android/helper/TestDriveUtils.kt
@@ -6,7 +6,7 @@ import androidx.compose.animation.core.SpringSpec
 import androidx.compose.animation.core.spring
 import androidx.compose.animation.core.tween
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material.Surface
+import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.platform.LocalDensity
@@ -14,11 +14,11 @@ import androidx.compose.ui.test.junit4.ComposeContentTestRule
 import com.bumble.appyx.components.internal.testdrive.TestDrive
 import com.bumble.appyx.components.internal.testdrive.TestDriveModel
 import com.bumble.appyx.components.internal.testdrive.TestDriveUi
-import com.bumble.appyx.components.internal.testdrive.ui.simple.TestDriveSimpleMotionController
-import com.bumble.appyx.interactions.core.gesture.GestureValidator.Companion.permissiveValidator
-import com.bumble.appyx.interactions.core.ui.helper.AppyxComponentSetup
-import com.bumble.appyx.interactions.sample.InteractionTarget
-import com.bumble.appyx.interactions.theme.appyx_dark
+import com.bumble.appyx.components.internal.testdrive.ui.simple.TestDriveSimpleVisualisation
+import com.bumble.appyx.interactions.gesture.GestureValidator.Companion.permissiveValidator
+import com.bumble.appyx.interactions.ui.helper.AppyxComponentSetup
+import com.bumble.appyx.interactions.utils.testing.TestTarget
+import com.bumble.appyx.interactions.utils.ui.theme.appyx_dark
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlin.math.roundToInt
@@ -30,23 +30,23 @@ fun ComposeContentTestRule.createTestDrive(
         easing = LinearEasing
     ),
     uiAnimationSpec: SpringSpec<Float> = spring()
-): TestDrive<InteractionTarget> {
+): TestDrive<TestTarget> {
     val model = TestDriveModel(
-        element = InteractionTarget.Child1,
+        element = TestTarget.Child1,
         savedStateMap = null
     )
     return TestDrive(
         scope = CoroutineScope(Dispatchers.Unconfined),
         model = model,
-        motionController = {
-            TestDriveSimpleMotionController(
+        visualisation = {
+            TestDriveSimpleVisualisation(
                 uiContext = it,
                 uiAnimationSpec = uiAnimationSpec
             )
         },
         progressAnimationSpec = animationSpec ?: spring(),
         gestureFactory = {
-            TestDriveSimpleMotionController.Gestures(it)
+            TestDriveSimpleVisualisation.Gestures(it)
         },
     ).also { setupTestDrive(it, model) }
 }