+++ demos/appyx-interactions/desktop/src/desktopMain/kotlin/com/bumble/appyx/interactions/widgets/ui/TargetUiState.kt
@@ -4,15 +4,15 @@ import androidx.compose.ui.graphics.TransformOrigin
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.dp
-import com.bumble.appyx.interactions.core.ui.context.UiContext
-import com.bumble.appyx.interactions.core.ui.math.clamp
-import com.bumble.appyx.interactions.core.ui.math.smoothstep
-import com.bumble.appyx.interactions.core.ui.property.impl.Alpha
-import com.bumble.appyx.interactions.core.ui.property.impl.position.PositionOutside
-import com.bumble.appyx.interactions.core.ui.property.impl.RotationX
-import com.bumble.appyx.interactions.core.ui.property.impl.Scale
-import com.bumble.appyx.interactions.core.ui.property.impl.ZIndex
-import com.bumble.appyx.interactions.core.ui.state.MutableUiStateSpecs
+import com.bumble.appyx.interactions.ui.context.UiContext
+import com.bumble.appyx.interactions.ui.math.clamp
+import com.bumble.appyx.interactions.ui.math.smoothstep
+import com.bumble.appyx.interactions.ui.property.impl.Alpha
+import com.bumble.appyx.interactions.ui.property.impl.RotationX
+import com.bumble.appyx.interactions.ui.property.impl.Scale
+import com.bumble.appyx.interactions.ui.property.impl.ZIndex
+import com.bumble.appyx.interactions.ui.property.impl.position.PositionOffset
+import com.bumble.appyx.interactions.ui.state.MutableUiStateSpecs
 import com.bumble.appyx.mapState
 import kotlinx.coroutines.flow.StateFlow
 
@@ -20,7 +20,7 @@ import kotlinx.coroutines.flow.StateFlow
 class TargetUiState(
     val positionInList: Int = 0,
     val rotationX: RotationX.Target,
-    val position: PositionOutside.Target,
+    val positionOffset: PositionOffset.Target,
     val scale: Scale.Target,
     val alpha: Alpha.Target,
     val zIndex: ZIndex.Target,
@@ -31,13 +31,13 @@ class TargetUiState(
     ) : this(
         positionInList = positionInList,
         rotationX = base.rotationX,
-        position = base.position,
+        positionOffset = base.positionOffset,
         scale = base.scale,
         alpha = base.alpha,
         zIndex = base.zIndex,
     )
 
-    fun toMutableState(
+    fun toMutableUiState(
         uiContext: UiContext,
         scrollX: StateFlow<Float>,
         itemWidth: Dp,
@@ -54,12 +54,12 @@ class TargetUiState(
                 },
                 origin = TransformOrigin(0.075f, 0f),
             ),
-            position = PositionOutside(
+            positionOffset = PositionOffset(
                 coroutineScope = uiContext.coroutineScope,
-                target = position,
+                target = positionOffset,
                 displacement = scrollX.mapState(uiContext.coroutineScope) {
                     val factor = 0.075f + smoothstep(0f, 1f, it)
-                    PositionOutside.Value(
+                    PositionOffset.Value(
                         offset = DpOffset(
                             x = (-0.125f * it * itemWidth.value).dp,
                             y = (-factor * it * itemHeight.value / (1f - 0.1f * it)).dp,