+++ demos/appyx-interactions/web/src/jsMain/kotlin/com/bumble/appyx/interactions/main.js.kt
@@ -7,11 +7,12 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
+import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.onSizeChanged
 import androidx.compose.ui.unit.IntSize
-import androidx.compose.ui.window.Window
+import androidx.compose.ui.window.CanvasBasedWindow
 import com.bumble.appyx.components.internal.testdrive.TestDriveExperiment
 import com.bumble.appyx.components.internal.testdrive.ui.md_amber_500
 import com.bumble.appyx.components.internal.testdrive.ui.md_blue_500
@@ -53,9 +54,10 @@ enum class InteractionTarget {
     Child1
 }
 
+@OptIn(ExperimentalComposeUiApi::class)
 fun main() {
     onWasmReady {
-        Window("Appyx") {
+        CanvasBasedWindow("Appyx") {
             var size by remember { mutableStateOf(IntSize.Zero) }
             Surface(
                 modifier = Modifier